Split off tests for duplicates into separate file
